Port CUDA extension fixes
