+++ src/ReinforcementLearningCore/test/runtests.jl
@@ -4,10 +4,18 @@ using ReinforcementLearningEnvironments
 using ReinforcementLearningTrajectories
 
 using Test
-using CUDA
 using CircularArrayBuffers
 using Flux
 
+@testset "ReinforcementLearningCore.jl" begin
+    include("core/core.jl")
+    include("core/stop_conditions.jl")
+    include("policies/policies.jl")
+    include("utils/utils.jl")
+end
+
+# Test again, with CUDA package extension loaded
+using CUDA
 @testset "ReinforcementLearningCore.jl" begin
     include("core/core.jl")
     include("core/stop_conditions.jl")
