+++ src/ReinforcementLearningCore/test/utils/distributions.jl
@@ -1,4 +1,4 @@
-using Test, LinearAlgebra, Distributions, ReinforcementLearningCore, CUDA, Flux
+using Test, LinearAlgebra, Distributions, ReinforcementLearningCore, Flux
 
 @testset "utils/distributions" begin
     @testset "logdetLorU" begin
@@ -8,7 +8,7 @@ using Test, LinearAlgebra, Distributions, ReinforcementLearningCore, CUDA, Flux
         logdetM = logdet(M)
         @test logdetM == ReinforcementLearningCore.logdetLorU(L) 
         @test logdetM == ReinforcementLearningCore.logdetLorU(U)
-        if CUDA.functional()
+        if (@isdefined CUDA) && CUDA.functional()
             L_d = cu(L)
             U_d = cu(U)
             @test logdetM == ReinforcementLearningCore.logdetLorU(L_d) 
@@ -126,7 +126,7 @@ using Test, LinearAlgebra, Distributions, ReinforcementLearningCore, CUDA, Flux
         end
     end
     @testset "CUDA" begin
-        if CUDA.functional()
+        if (@isdefined CUDA) && CUDA.functional()
             CUDA.allowscalar(false)
             #These only test GPU compatibility, exactness of results is tested above on the CPU
             @testset "Diagonal Gaussian" begin
